+++ src/main/java/team5/dao/BigliettoDAO.java
@@ -68,16 +68,16 @@ public class BigliettoDAO {
     }
 
     //  numero biglietti vidimati su un mezzo
-    public long numeroBigliettiVidimatiPerMezzo(Mezzo mezzo) {
-        TypedQuery<Long> query = entityManager.createQuery("SELECT COUNT(b) FROM Biglietto b WHERE b.data_validazione IS NOT NULL AND b.mezzi = :mezzo", Long.class);
+    public int numeroBigliettiVidimatiPerMezzo(Mezzo mezzo) {
+        TypedQuery<Integer> query = entityManager.createQuery("SELECT COUNT(b) FROM Biglietto b WHERE b.data_validazione IS NOT NULL AND b.mezzi = :mezzo", Integer.class);
         query.setParameter("mezzo", mezzo);
         return query.getSingleResult();
     }
 
     // numero biglietti vidimati in un periodo
-    public long numeroBigliettiVidimatiDaData(LocalDate inizio) {
-        TypedQuery<Long> query = entityManager.createQuery(
-                "SELECT COUNT(b) FROM Biglietto b WHERE b.data_validazione IS NOT NULL AND b.data_validazione >= :inizio", Long.class);
+    public int numeroBigliettiVidimatiDaData(LocalDate inizio) {
+        TypedQuery<Integer> query = entityManager.createQuery(
+                "SELECT COUNT(b) FROM Biglietto b WHERE b.data_validazione IS NOT NULL AND b.data_validazione >= :inizio", Integer.class);
 
         query.setParameter("inizio", inizio);
         return query.getSingleResult();
@@ -95,7 +95,7 @@ public class BigliettoDAO {
             biglietto.setRivenditore(rivenditore);
 
             //GENERA UN CODICE UNIVOCO
-            Long count = (Long) entityManager.createQuery("SELECT COUNT (b) FROM Biglietto b").getSingleResult();
+            int count = (Integer) entityManager.createQuery("SELECT COUNT (b) FROM Biglietto b").getSingleResult();
             String codice = String.format("B-%04d", count + 1);
             biglietto.setCodice_univoco(codice);
 
@@ -111,10 +111,10 @@ public class BigliettoDAO {
     }
 
     //NUMERO DI BIGLIETTI EMESSI DA UN PUNTO VENDITA E PER UN PERIODO DI TEMPO
-    public Long numeroBigliettiEmessiPerRivenditoriEPerPeriodo(Rivenditore rivenditore, LocalDate inizio, LocalDate fine) {
-        TypedQuery<Long> query = entityManager.createQuery("" +
+    public int numeroBigliettiEmessiPerRivenditoriEPerPeriodo(Rivenditore rivenditore, LocalDate inizio, LocalDate fine) {
+        TypedQuery<Integer> query = entityManager.createQuery("" +
                 "SELECT COUNT(b) FROM Biglietto b " +
-                "WHERE b.rivenditore = :rivenditore AND b.data_emissione BETWEEN :inizio AND :fine", Long.class);
+                "WHERE b.rivenditore = :rivenditore AND b.data_emissione BETWEEN :inizio AND :fine", Integer.class);
         query.setParameter("rivenditore", rivenditore);
         query.setParameter("inizio", inizio);
         query.setParameter("fine", fine);
